+++ tests/base/eval_model_template.py
@@ -2,18 +2,18 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from tests.base.config_optimizers import ConfigureOptimizersPool
 from tests.base.datasets import TrialMNIST
 from pytorch_lightning.core.lightning import LightningModule
-from tests.base.model_test_dataloaders import TestDataloaderVariations
-from tests.base.model_test_steps import TestStepVariations
-from tests.base.model_train_dataloaders import TrainDataloaderVariations
-from tests.base.model_train_steps import TrainingStepVariations
-from tests.base.model_utils import ModelTemplateUtils
-from tests.base.model_valid_dataloaders import ValDataloaderVariations
-from tests.base.model_valid_steps import ValidationStepVariations
-from tests.base.trainer_test_epoch_ends import TestEpochEndVariations
-from tests.base.trainer_valid_epoch_ends import ValidationEpochEndVariations
+from tests.base.eval_model_optimizers import ConfigureOptimizersPool
+from tests.base.eval_model_test_dataloaders import TestDataloaderVariations
+from tests.base.eval_model_test_epoch_ends import TestEpochEndVariations
+from tests.base.eval_model_test_steps import TestStepVariations
+from tests.base.eval_model_train_dataloaders import TrainDataloaderVariations
+from tests.base.eval_model_train_steps import TrainingStepVariations
+from tests.base.eval_model_valid_dataloaders import ValDataloaderVariations
+from tests.base.eval_model_valid_epoch_ends import ValidationEpochEndVariations
+from tests.base.eval_model_valid_steps import ValidationStepVariations
+from tests.base.eval_model_utils import ModelTemplateUtils
 
 
 class ModelTrailTemplate(
